+++ testing/test_full.py
@@ -69,6 +69,7 @@ def test_train_segmentation_rad_2d():
   print('Starting 2D Rad segmentation tests')
   application_data = '2d_rad_segmentation'
   parameters = parseConfig(inputDir + '/' + application_data + '/sample_training.yaml', version_check = False)
+  parameters['modality'] = 'rad'
   training_data, headers = parseTrainingCSV(inputDir + '/train_' + application_data + '.csv')
   for model in all_models_segmentation:
     parameters['model']['architecture'] = model 
@@ -82,6 +83,7 @@ def test_train_segmentation_rad_3d():
   print('Starting 3D Rad segmentation tests')
   application_data = '3d_rad_segmentation'
   parameters = parseConfig(inputDir + '/' + application_data + '/sample_training.yaml', version_check = False)
+  parameters['modality'] = 'rad'
   training_data, headers = parseTrainingCSV(inputDir + '/train_' + application_data + '.csv')
   for model in all_models_segmentation:
     parameters['model']['architecture'] = model